:white_check_mark: Move unit tests and fix
